requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bitsandbytes==0.43.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.24.6
idna==3.8
importlib_resources==6.4.4
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
latex2mathml==3.77.0
-e git+https://github.com/He-Changhao/LLaVA-VideoDigest@3382b47dafb0412e231ea564a7e51805c75f4e0d#egg=llava
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.5.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
protobuf==5.27.3
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
ruff==0.6.2
safetensors==0.4.4
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
svgwrite==1.4.3
sympy==1.13.2
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.37.2
triton==2.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
